requirements.txt
streamlit
requests
beautifulsoup4
google-genai
python-dotenv
pandas
